LayoutTests/fast/dom/Element/id-in-deletebutton.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<div contenteditable="true">
<ul class="needsDeletionUI"><li>1</li><li id="li">2</li></ul>

<script>
description('Test document.getElementById("WebKit-Editing-Delete-Button")');

sel = window.getSelection();
li = document.getElementById("li");
sel.setPosition(li, 0);

if (window.testRunner) {
    deleteButton = document.getElementById("WebKit-Editing-Delete-Button");
    shouldBeNonNull('deleteButton');
}
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
